+++ Tests/GRDBTests/FTS5TokenizerTests.swift
@@ -225,5 +225,85 @@ class FTS5TokenizerTests: GRDBTestCase {
             XCTAssertFalse(match(db, "2016-10-04.txt", "txt"))
         }
     }
+    
+    func testTokenize() throws {
+        try makeDatabaseQueue().inDatabase { db in
+            let ascii = try db.makeTokenizer(.ascii())
+            let porter = try db.makeTokenizer(.porter())
+            let unicode61 = try db.makeTokenizer(.unicode61())
+            let unicode61WithDiacritics = try db.makeTokenizer(.unicode61(diacritics: .keep))
+
+            // Empty query
+            try XCTAssertEqual(ascii.tokenize(query: "").map(\.token), [])
+            try XCTAssertEqual(porter.tokenize(query: "").map(\.token), [])
+            try XCTAssertEqual(unicode61.tokenize(query: "").map(\.token), [])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "").map(\.token), [])
+            
+            try XCTAssertEqual(ascii.tokenize(query: "?!").map(\.token), [])
+            try XCTAssertEqual(porter.tokenize(query: "?!").map(\.token), [])
+            try XCTAssertEqual(unicode61.tokenize(query: "?!").map(\.token), [])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "?!").map(\.token), [])
+            
+            // Token queries
+            try XCTAssertEqual(ascii.tokenize(query: "Moby").map(\.token), ["moby"])
+            try XCTAssertEqual(porter.tokenize(query: "Moby").map(\.token), ["mobi"])
+            try XCTAssertEqual(unicode61.tokenize(query: "Moby").map(\.token), ["moby"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "Moby").map(\.token), ["moby"])
+            
+            try XCTAssertEqual(ascii.tokenize(query: "écarlates").map(\.token), ["écarlates"])
+            try XCTAssertEqual(porter.tokenize(query: "écarlates").map(\.token), ["ecarl"])
+            try XCTAssertEqual(unicode61.tokenize(query: "écarlates").map(\.token), ["ecarlates"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "écarlates").map(\.token), ["écarlates"])
+            
+            try XCTAssertEqual(ascii.tokenize(query: "fooéı👨👨🏿🇫🇷🇨🇮").map(\.token), ["fooéı👨👨🏿🇫🇷🇨🇮"])
+            try XCTAssertEqual(porter.tokenize(query: "fooéı👨👨🏿🇫🇷🇨🇮").map(\.token), ["fooeı", "🏿"]) // ¯\_(ツ)_/¯
+            try XCTAssertEqual(unicode61.tokenize(query: "fooéı👨👨🏿🇫🇷🇨🇮").map(\.token), ["fooeı", "🏿"]) // ¯\_(ツ)_/¯
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "fooéı👨👨🏿🇫🇷🇨🇮").map(\.token), ["fooéı", "🏿"]) // ¯\_(ツ)_/¯
+            
+            try XCTAssertEqual(ascii.tokenize(query: "SQLite database").map(\.token), ["sqlite", "database"])
+            try XCTAssertEqual(porter.tokenize(query: "SQLite database").map(\.token), ["sqlite", "databas"])
+            try XCTAssertEqual(unicode61.tokenize(query: "SQLite database").map(\.token), ["sqlite", "database"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "SQLite database").map(\.token), ["sqlite", "database"])
+            
+            try XCTAssertEqual(ascii.tokenize(query: "Édouard Manet").map(\.token), ["Édouard", "manet"])
+            try XCTAssertEqual(porter.tokenize(query: "Édouard Manet").map(\.token), ["edouard", "manet"])
+            try XCTAssertEqual(unicode61.tokenize(query: "Édouard Manet").map(\.token), ["edouard", "manet"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "Édouard Manet").map(\.token), ["édouard", "manet"])
+            
+            // Prefix queries
+            try XCTAssertEqual(ascii.tokenize(query: "*").map(\.token), [])
+            try XCTAssertEqual(porter.tokenize(query: "*").map(\.token), [])
+            try XCTAssertEqual(unicode61.tokenize(query: "*").map(\.token), [])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "*").map(\.token), [])
+            
+            try XCTAssertEqual(ascii.tokenize(query: "Robin*").map(\.token), ["robin"])
+            try XCTAssertEqual(porter.tokenize(query: "Robin*").map(\.token), ["robin"])
+            try XCTAssertEqual(unicode61.tokenize(query: "Robin*").map(\.token), ["robin"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "Robin*").map(\.token), ["robin"])
+            
+            // Phrase queries
+            try XCTAssertEqual(ascii.tokenize(query: "\"foulent muscles\"").map(\.token), ["foulent", "muscles"])
+            try XCTAssertEqual(porter.tokenize(query: "\"foulent muscles\"").map(\.token), ["foulent", "muscl"])
+            try XCTAssertEqual(unicode61.tokenize(query: "\"foulent muscles\"").map(\.token), ["foulent", "muscles"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "\"foulent muscles\"").map(\.token), ["foulent", "muscles"])
+            
+            try XCTAssertEqual(ascii.tokenize(query: "\"Kim Stan* Robin*\"").map(\.token), ["kim", "stan", "robin"])
+            try XCTAssertEqual(porter.tokenize(query: "\"Kim Stan* Robin*\"").map(\.token), ["kim", "stan", "robin"])
+            try XCTAssertEqual(unicode61.tokenize(query: "\"Kim Stan* Robin*\"").map(\.token), ["kim", "stan", "robin"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "\"Kim Stan* Robin*\"").map(\.token), ["kim", "stan", "robin"])
+            
+            // Logical queries
+            try XCTAssertEqual(ascii.tokenize(query: "years AND months").map(\.token), ["years", "and", "months"])
+            try XCTAssertEqual(porter.tokenize(query: "years AND months").map(\.token), ["year", "and", "month"])
+            try XCTAssertEqual(unicode61.tokenize(query: "years AND months").map(\.token), ["years", "and", "months"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "years AND months").map(\.token), ["years", "and", "months"])
+            
+            // column queries
+            try XCTAssertEqual(ascii.tokenize(query: "title:brest").map(\.token), ["title", "brest"])
+            try XCTAssertEqual(porter.tokenize(query: "title:brest").map(\.token), ["titl", "brest"])
+            try XCTAssertEqual(unicode61.tokenize(query: "title:brest").map(\.token), ["title", "brest"])
+            try XCTAssertEqual(unicode61WithDiacritics.tokenize(query: "title:brest").map(\.token), ["title", "brest"])
+        }
+    }
 }
 #endif